chore(test): set up testing framework with initial examples
chore(test): set up snapshot test helper to compare screenshots of different stages of tests with previous successful tests using hashes

+++ ubo_app/side_effects.py
@@ -3,14 +3,17 @@ from __future__ import annotations
 
 import atexit
 import subprocess
+import weakref
 from typing import TYPE_CHECKING, Sequence
 
 from debouncer import DebounceOptions, debounce
-from kivy.clock import Clock
-from redux import FinishAction, FinishEvent
+from kivy.clock import mainthread
+from redux import AutorunOptions, EventSubscriptionOptions, FinishAction, FinishEvent
 
 from ubo_app.store import autorun, dispatch, subscribe_event
 from ubo_app.store.main import PowerOffEvent
+from ubo_app.store.services.notifications import Chime
+from ubo_app.store.services.sound import SoundPlayChimeAction
 from ubo_app.store.update_manager import (
     UpdateManagerCheckEvent,
     UpdateManagerSetStatusAction,
@@ -20,16 +23,22 @@ from ubo_app.store.update_manager import (
 from ubo_app.store.update_manager.reducer import ABOUT_MENU_PATH
 from ubo_app.store.update_manager.utils import check_version, update
 from ubo_app.utils.async_ import create_task
-from ubo_app.utils.hardware import initialize_board, turn_off_screen, turn_on_screen
+from ubo_app.utils.hardware import (
+    IS_RPI,
+    initialize_board,
+    turn_off_screen,
+    turn_on_screen,
+)
 
 if TYPE_CHECKING:
     from ubo_app.menu import MenuApp
 
 
-def power_off(_: PowerOffEvent) -> None:
+def power_off() -> None:
     """Power off the device."""
-    dispatch(FinishAction())
-    subprocess.run(['/usr/bin/env', 'systemctl', 'poweroff', '-i'], check=True)  # noqa: S603
+    dispatch(SoundPlayChimeAction(name=Chime.FAILURE), FinishAction())
+    if IS_RPI:
+        subprocess.run(['/usr/bin/env', 'systemctl', 'poweroff', '-i'], check=True)  # noqa: S603
 
 
 def setup(app: MenuApp) -> None:
@@ -37,10 +46,35 @@ def setup(app: MenuApp) -> None:
     turn_on_screen()
     initialize_board()
 
-    subscribe_event(PowerOffEvent, power_off)
-    subscribe_event(FinishEvent, lambda *_: Clock.schedule_once(app.stop))
-    subscribe_event(UpdateManagerUpdateEvent, lambda: create_task(update()))
-    subscribe_event(UpdateManagerCheckEvent, lambda: create_task(check_version()))
+    subscribe_event(
+        PowerOffEvent,
+        power_off,
+        options=EventSubscriptionOptions(keep_ref=False),
+    )
+
+    app_ref = weakref.ref(app)
+
+    @mainthread
+    def stop_app() -> None:
+        app = app_ref()
+        if app is not None:
+            app.stop()
+
+    subscribe_event(
+        FinishEvent,
+        stop_app,
+        options=EventSubscriptionOptions(keep_ref=True),
+    )
+    subscribe_event(
+        UpdateManagerUpdateEvent,
+        update,
+        options=EventSubscriptionOptions(keep_ref=False),
+    )
+    subscribe_event(
+        UpdateManagerCheckEvent,
+        check_version,
+        options=EventSubscriptionOptions(keep_ref=False),
+    )
 
     @debounce(
         wait=10,
@@ -49,8 +83,8 @@ def setup(app: MenuApp) -> None:
     async def request_check_version() -> None:
         dispatch(UpdateManagerSetStatusAction(status=UpdateStatus.CHECKING))
 
-    @autorun(lambda state: state.main.path)
-    def check_version_caller(
+    @autorun(lambda state: state.main.path, options=AutorunOptions(keep_ref=False))
+    def _(
         path: Sequence[str],
         previous_path: Sequence[str],
     ) -> None:
